KidneyStoneDetection_YOLOv8_Project
This repository houses my project focused on artificial intelligence, specifically utilizing YOLOv8 for the detection of kidney stones in X-ray images. Explore the implementation and results of an intelligent system designed to identify and locate kidney stones with high precision, contributing to the field of medical image analysis and healthcare innovation.