LunarMartianCraterDetection_Project_Description
This document provides an overview of our project, where we tackled the exciting challenge of detecting craters in images of lunar and Martian surfaces. Utilizing the open-source TorchVision library and our profound interest in space exploration, we embarked on this journey to develop a precise and efficient crater detection system.